#669 refactor to make layout generation optional
