+++ xpath-outcome-initiator/src/test/data/outcomeBuilder/dynamicForms/EmployeeShiftSchedule.xsd
@@ -1,0 +1,75 @@
+<?xml version='1.0' encoding='utf-8'?>
+<!--
+
+    This file is part of the CRISTAL-iSE XPath Outcome Initiator module.
+    Copyright (c) 2001-2016 The CRISTAL Consortium. All rights reserved.
+
+    This library is free software; you can redistribute it and/or modify it
+    under the terms of the GNU Lesser General Public License as published
+    by the Free Software Foundation; either version 3 of the License, or (at
+    your option) any later version.
+
+    This library is distributed in the hope that it will be useful, but WITHOUT
+    ANY WARRANTY; with out even the implied warranty of MERCHANTABILITY or
+    FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public
+    License for more details.
+
+    You should have received a copy of the GNU Lesser General Public License
+    along with this library; if not, write to the Free Software Foundation,
+    Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA.
+
+    http://www.fsf.org/licensing/licenses/lgpl.html
+
+-->
+<xs:schema xmlns:xs='http://www.w3.org/2001/XMLSchema'>
+  <xs:element name='EmployeeShiftSchedule'>
+    <xs:annotation>
+      <xs:documentation>The actual shift assigned with employee</xs:documentation>
+    </xs:annotation>
+    <xs:complexType>
+      <xs:all minOccurs='0'>
+        <xs:element name='ShiftName' type='xs:string' minOccurs='1' maxOccurs='1'>
+          <xs:annotation>
+            <xs:appinfo>
+              <listOfValues>
+                <inputName>memberNames</inputName>
+              </listOfValues>
+            </xs:appinfo>
+          </xs:annotation>
+        </xs:element>
+        <xs:element name='CollectionName' type='xs:string' minOccurs='1' maxOccurs='1'>
+          <xs:annotation>
+            <xs:appinfo>
+              <dynamicForms>
+                <required>false</required>
+              </dynamicForms>
+              <listOfValues>
+                <values>Billing,Shipping,Head Office</values>
+              </listOfValues>
+            </xs:appinfo>
+          </xs:annotation>
+        </xs:element>
+        <xs:element name='MemberUUID' type='xs:string' minOccurs='1' maxOccurs='1'>
+          <xs:annotation>
+            <xs:appinfo>
+              <dynamicForms>
+                <hidden>true</hidden>
+                <required>false</required>
+              </dynamicForms>
+            </xs:appinfo>
+          </xs:annotation>
+        </xs:element>
+        <xs:element name='MemberID' type='xs:integer' minOccurs='1' maxOccurs='1'>
+          <xs:annotation>
+            <xs:appinfo>
+              <dynamicForms>
+                <hidden>true</hidden>
+                <required>false</required>
+              </dynamicForms>
+            </xs:appinfo>
+          </xs:annotation>
+        </xs:element>
+      </xs:all>
+    </xs:complexType>
+  </xs:element>
+</xs:schema>
