+++ xpath-outcome-initiator/src/test/java/org/cristalise/kernel/test/persistency/outcomebuilder/ListOfValuesTest.java
@@ -34,13 +34,12 @@ import org.cristalise.kernel.persistency.outcomebuilder.OutcomeBuilder;
 import org.cristalise.kernel.process.Gateway;
 import org.cristalise.kernel.test.persistency.XMLUtils;
 import org.json.JSONArray;
-import org.json.JSONObject;
 import org.junit.Before;
 import org.junit.Test;
 
 public class ListOfValuesTest extends XMLUtils {
 
-    String dir = "src/test/data/outcomeBuilder";
+    String dir = "src/test/data/outcomeBuilder/dynamicForms";
 
     @Before
     public void setUp() throws Exception {
@@ -82,15 +81,5 @@ public class ListOfValuesTest extends XMLUtils {
         JSONArray expected = new JSONArray(getJSON(dir, type+"_emptyInputs"));
 
         assertJsonEquals(expected, actual);
-    }
-
-    @Test
-    public void employeeShiftScheduleFromJson() throws Exception {
-        String type = "EmployeeShiftSchedule";
-        OutcomeBuilder builder = new OutcomeBuilder(new Schema(type, 0, getXSD(dir, type)), true);
-
-        builder.addJsonInstance(new JSONObject("{'EmployeeShiftSchedule': {'CollectionName': 'Shift','MemberID': '0','ShiftName': 'shift1','MemberUUID': null,}}"));
-        String actual = builder.getXml();
-        assert compareXML(getXML(dir, type), actual);
     }
 }
